+++ lib/project_types/theme/messages/messages.rb
@@ -65,6 +65,8 @@ module Theme
                 {{command:-j, --json}}             Output JSON instead of a UI.
                 {{command:-a, --allow-live}}       Allow push to a live theme.
                 {{command:-p, --publish}}          Publish as the live theme after uploading.
+                {{command:-o, --only}}             Upload only the specified files.
+                {{command:-x, --ignore}}           Skip uploading the specified files.
 
               Run without options to select theme from a list.
           HELP
@@ -200,6 +202,8 @@ module Theme
               {{command:-l, --live}}             Pull theme files from your remote live theme.
               {{command:-d, --development}}      Pull theme files from your remote development theme.
               {{command:-n, --nodelete}}         Runs the pull command without deleting local files.
+              {{command:-o, --only}}             Download only the specified files.
+              {{command:-x, --ignore}}           Skip downloading the specified files.
 
             Run without options to select theme from a list.
           HELP