+++ test/project_types/theme/commands/pull_test.rb
@@ -20,6 +20,10 @@ module Theme
         )
         @syncer = stub("Syncer", lock_io!: nil, unlock_io!: nil)
         @ignore_filter = mock("IgnoreFilter")
+        @include_filter = mock("IncludeFilter")
+
+        ShopifyCLI::Theme::IgnoreFilter.stubs(:from_path).with(".").returns(@ignore_filter)
+        ShopifyCLI::Theme::IncludeFilter.stubs(:new).returns(@include_filter)
       end
 
       def test_pull_with_deprecated_theme_id
@@ -27,10 +31,8 @@ module Theme
           .with(@ctx, root: ".", id: 1234)
           .returns(@theme)
 
-        ShopifyCLI::Theme::IgnoreFilter.expects(:from_path).with(".").returns(@ignore_filter)
-
         ShopifyCLI::Theme::Syncer.expects(:new)
-          .with(@ctx, theme: @theme, ignore_filter: @ignore_filter)
+          .with(@ctx, theme: @theme, include_filter: @include_filter, ignore_filter: @ignore_filter)
           .returns(@syncer)
 
         @syncer.expects(:start_threads)
@@ -52,7 +54,7 @@ module Theme
         ShopifyCLI::Theme::IgnoreFilter.expects(:from_path).with(".").returns(@ignore_filter)
 
         ShopifyCLI::Theme::Syncer.expects(:new)
-          .with(@ctx, theme: @theme, ignore_filter: @ignore_filter)
+          .with(@ctx, theme: @theme, include_filter: @include_filter, ignore_filter: @ignore_filter)
           .returns(@syncer)
 
         @syncer.expects(:start_threads)
@@ -73,7 +75,7 @@ module Theme
         ShopifyCLI::Theme::IgnoreFilter.expects(:from_path).with(".").returns(@ignore_filter)
 
         ShopifyCLI::Theme::Syncer.expects(:new)
-          .with(@ctx, theme: @theme, ignore_filter: @ignore_filter)
+          .with(@ctx, theme: @theme, include_filter: @include_filter, ignore_filter: @ignore_filter)
           .returns(@syncer)
 
         @syncer.expects(:start_threads)
@@ -91,10 +93,8 @@ module Theme
           .with(@ctx, root: ".")
           .returns(@theme)
 
-        ShopifyCLI::Theme::IgnoreFilter.expects(:from_path).with(".").returns(@ignore_filter)
-
         ShopifyCLI::Theme::Syncer.expects(:new)
-          .with(@ctx, theme: @theme, ignore_filter: @ignore_filter)
+          .with(@ctx, theme: @theme, include_filter: @include_filter, ignore_filter: @ignore_filter)
           .returns(@syncer)
 
         @syncer.expects(:start_threads)
@@ -115,7 +115,7 @@ module Theme
         ShopifyCLI::Theme::IgnoreFilter.expects(:from_path).with(".").returns(@ignore_filter)
 
         ShopifyCLI::Theme::Syncer.expects(:new)
-          .with(@ctx, theme: @theme, ignore_filter: @ignore_filter)
+          .with(@ctx, theme: @theme, include_filter: @include_filter, ignore_filter: @ignore_filter)
           .returns(@syncer)
 
         @syncer.expects(:start_threads)
@@ -133,10 +133,8 @@ module Theme
           .with(@ctx, root: ".", identifier: 1234)
           .returns(@theme)
 
-        ShopifyCLI::Theme::IgnoreFilter.expects(:from_path).with(".").returns(@ignore_filter)
-
         ShopifyCLI::Theme::Syncer.expects(:new)
-          .with(@ctx, theme: @theme, ignore_filter: @ignore_filter)
+          .with(@ctx, theme: @theme, include_filter: @include_filter, ignore_filter: @ignore_filter)
           .returns(@syncer)
 
         @syncer.expects(:start_threads)
@@ -151,16 +149,43 @@ module Theme
         @command.call([], "pull")
       end
 
+      def test_pull_with_filter
+        includes = "config/*"
+        include_filter = mock("IncludeFilter")
+
+        ShopifyCLI::Theme::Theme.expects(:find_by_identifier)
+          .with(@ctx, root: ".", identifier: 1234)
+          .returns(@theme)
+        ShopifyCLI::Theme::IncludeFilter.expects(:new)
+          .with(includes)
+          .returns(include_filter)
+        ShopifyCLI::Theme::Syncer.expects(:new)
+          .with(@ctx, theme: @theme, include_filter: include_filter, ignore_filter: @ignore_filter)
+          .returns(@syncer)
+
+        @syncer.expects(:start_threads)
+        @syncer.expects(:shutdown)
+
+        @syncer.expects(:download_theme!).with(delete: true)
+        @ctx.expects(:done)
+
+        @command.options.flags[:theme] = 1234
+        @command.options.flags[:includes] = includes
+        @command.call([], "pull")
+      end
+
       def test_pull_with_ignores
         ShopifyCLI::Theme::Theme.expects(:find_by_identifier)
           .with(@ctx, root: ".", identifier: 1234)
           .returns(@theme)
 
-        ShopifyCLI::Theme::IgnoreFilter.expects(:from_path).with(".").returns(@ignore_filter)
         @ignore_filter.expects(:add_patterns).with(["config/*"])
 
+        ShopifyCLI::Theme::IgnoreFilter.expects(:from_path)
+          .with(".")
+          .returns(@ignore_filter)
         ShopifyCLI::Theme::Syncer.expects(:new)
-          .with(@ctx, theme: @theme, ignore_filter: @ignore_filter)
+          .with(@ctx, theme: @theme, include_filter: @include_filter, ignore_filter: @ignore_filter)
           .returns(@syncer)
 
         @syncer.expects(:start_threads)
@@ -181,10 +206,8 @@ module Theme
 
         @syncer.expects(:download_theme!).with(delete: true)
 
-        ShopifyCLI::Theme::IgnoreFilter.expects(:from_path).with(".").returns(@ignore_filter)
-
         ShopifyCLI::Theme::Syncer.expects(:new)
-          .with(@ctx, theme: @theme, ignore_filter: @ignore_filter)
+          .with(@ctx, theme: @theme, include_filter: @include_filter, ignore_filter: @ignore_filter)
           .returns(@syncer)
 
         @syncer.expects(:start_threads)