+++ test/shopify-cli/theme/syncer_test.rb
@@ -256,6 +256,34 @@ module ShopifyCLI
         assert_empty(@syncer)
       end
 
+      def test_download_theme_with_include_filter
+        @syncer.include_filter = mock("IncludeFilter")
+        @syncer.include_filter.stubs(:match?).returns(false)
+        @syncer.include_filter.expects(:match?).with("layout/theme.liquid").returns(true)
+
+        @syncer.start_threads
+        @syncer.checksums.replace(@theme.theme_files.map { |file| [file.relative_path.to_s, "OUTDATED"] }.to_h)
+
+        File.any_instance.expects(:write)
+          .with("new content")
+          .once
+
+        ShopifyCLI::AdminAPI.expects(:rest_request)
+          .times(2) # +1 for checksums
+          .returns([
+            200,
+            {
+              "asset" => {
+                "value" => "new content",
+              },
+            },
+            {},
+          ])
+
+        @syncer.download_theme!
+        assert_empty(@syncer)
+      end
+
       def test_download_theme_with_ignores
         @syncer.ignore_filter = mock("IgnoreFilter")
         @syncer.ignore_filter.stubs(:ignore?).returns(false)